alpaca_client/pyproject.toml
[tool.poetry]
name = "alpaca-client"
version = "0.1.0"
description = ""
authors = ["Ryan Anderson <ryan@anderson.vc>"]

[tool.poetry.dependencies]
python = "^3.9"
websocket-client = "^1.2.1"
python-dotenv = "^0.19.1"
websockets = "^10.0"
asyncio = "^3.4.3"
aiohttp = "^3.8.0"
psycopg2-binary = "^2.9.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
